+++ README.md
@@ -121,10 +121,9 @@ npm run openapi:generate
 │   ├── src/
 │   │   ├── app/            # App Router のページと layout
 │   │   ├── features/       # feature 単位で切り出した画面実装
-│   │   ├── application/    # feature 化前の画面ユースケースと DTO
+│   │   ├── application/    # 画面・通信 DTO
 │   │   ├── domain/         # 業務ルールと業務型
 │   │   ├── infrastructure/ # API / localStorage 実装
-│   │   ├── presentation/   # study-session など移行前の画面実装
 │   │   └── __tests__/      # フロントのユースケース / infrastructure テスト
 │   ├── public/
 │   ├── next.config.mjs
@@ -145,8 +144,9 @@ npm run openapi:generate
 - 短文クイズは後方互換のためアプリ内データとして配信する
 - バックエンドは `domain -> application -> infrastructure -> presentation` の依存方向で責務分離する
 - フロントエンドも `presentation -> application -> domain` を基本にし、`infrastructure` が API と `localStorage` の詳細を担当する
-- `home-dashboard` と `result-screen` は `frontend/src/features` 配下へ移し、`study-session` は段階的に追随させる
-- 移行前の画面は `frontend/src/presentation` と `frontend/src/application/usecases` に残し、小さい単位で feature 化を進める
+- `home-dashboard`、`result-screen`、`study-session` は `frontend/src/features` 配下へ移し、画面単位で責務を閉じ込める
+- タイピング中の入力判定は `frontend/src/features/study-session/typing` に寄せ、フロント固有ロジックとして扱う
+- `frontend/src/application` には DTO を置き、画面実装は `features` と `app` を中心に構成する
 - 業務型は [`frontend/src/domain/models/study.ts`](./frontend/src/domain/models/study.ts)、画面・通信 DTO は [`frontend/src/application/dtos/study.ts`](./frontend/src/application/dtos/study.ts) に分離する
 - `localStorage` には `mistake_log`、最新結果、復習用キュー、設定値を保存
 - SQLite は `DATABASE_URL` で切り替え可能で、`docker compose` では named volume 上の `sqlite:////data/app.db` を使用
