+++ frontend/src/features/study-session/ui/StudySession.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Link from "next/link";
-import { useStudySession } from "@/presentation/hooks/useStudySession";
+import { useStudySession } from "@/features/study-session/hooks/useStudySession";
 import type { StudyMode } from "@/domain/models/study";
 
 function formatMs(ms: number) {
